Refactored to use wp-scripts and a ton of other stuff
