--- assets/src/js/block/editor.scss
@@ -1,3 +1,0 @@
-/**
- * The following styles get applied inside the editor only.
- */
