+++ block.json
@@ -6,8 +6,9 @@
     "category": "widgets",
     "description": "Easily display Google Reviews on your WordPress website using a powerful and intuitive widget.",
     "textdomain": "google-places-reviews",
-    "editorScript": "file:./assets/dist/js/block.js",
-    "editorStyle": "file:./assets/dist/css/admin-main.css",
+    "editorScript": "file:./build/google-block.js",
+    "editorStyle": "file:./build/google-block.css",
+    "style": ["file:./build/style-google-block.css", "google-block-styles"],
     "attributes": {
         "placeId": {
             "type": "string",
