+++ includes/legacy/assets/css/public-widget.scss
