+++ includes/legacy/assets/images/glyphicons-halflings.png
