+++ includes/legacy/assets/images/google-logo-options.png
