+++ includes/legacy/assets/images/new-window-icon.png
