+++ includes/legacy/assets/images/platform-icon-wpbr.png
