+++ includes/legacy/assets/images/success.png
