+++ includes/legacy/assets/js/admin-widget.js
