--- postcss.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-	plugins: [
-		require( 'autoprefixer' )
-	]
-};
