+++ src/editor.scss
@@ -1,0 +1,38 @@
+/**
+ * The following styles get applied inside the editor only.
+ */
+
+// 👋 Welcome and API Key.
+#rbg-admin-welcome-wrap {
+    display: flex;
+    align-items: center;
+    text-align: center;
+    justify-content: center;
+    padding: 30px 30px 60px;
+
+    .rbg-admin-google-logo {
+        margin: 20px auto;
+        width: 300px;
+    }
+
+    #google-block-admin-lottie-api {
+        margin: -70px 0 10px;
+    }
+
+    .rbg-admin-welcome-content-wrap {
+        max-width: 380px;
+        margin: 0 auto 30px;
+    }
+
+    .rbg-admin-google-welcome-heading {
+        font-weight: 600;
+        font-size: 24px;
+        line-height: 30px;
+        margin: 0 0 15px;
+        padding: 0;
+    }
+
+    .rbg-admin-button {
+        margin: 20px 0 0;
+    }
+}
