+++ src/style.scss
@@ -1,0 +1,14 @@
+/**
+ * Applied to frontend and editor views.
+ */
+
+@import 'styles/variables';
+
+[id^="rbg-"] {
+    position: relative;
+    margin: 50px auto;
+    padding: 20px;
+    border-radius: $border-radius;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px, rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
+    line-height: 1.6;
+}
