+++ src/styles/variables.scss
@@ -1,0 +1,6 @@
+
+$border-color: #E1E4E8;
+$border-radius: 8px;
+$heading-color: #24292F;
+$link-color: #027A97;
+$yelp-color: #E00707;
